Events Added

+++ src/Event.java
@@ -1,5 +1,5 @@
 public interface Event{
 
     void doEvent();
-    float getTime();
+    double getTime();
 }
